Añadir crear, borrar y editar Notas y componentes
