+++ app/src/main/java/com/example/dispositivomobil/components/NoteItem.kt
@@ -1,0 +1,29 @@
+package com.example.dispositivomobil.components
+
+import androidx.compose.foundation.layout.*
+import androidx.compose.material3.*
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import com.example.dispositivomobil.models.Note  // Importa la clase Note
+
+@Composable
+fun NoteItem(note: Note, onEdit: () -> Unit, onDelete: () -> Unit) {
+    Card(modifier = Modifier.padding(8.dp), elevation = CardDefaults.cardElevation(4.dp)) {
+        Column(modifier = Modifier.padding(16.dp)) {
+            Text(text = note.title, style = MaterialTheme.typography.titleMedium)
+            Spacer(modifier = Modifier.height(4.dp))
+            Text(text = note.description)
+            Spacer(modifier = Modifier.height(8.dp))
+            Row {
+                Button(onClick = onEdit) {
+                    Text("Edit")
+                }
+                Spacer(modifier = Modifier.width(8.dp))
+                Button(onClick = onDelete) {
+                    Text("Delete")
+                }
+            }
+        }
+    }
+}
