+++ app/src/main/java/com/example/dispositivomobil/models/Note.kt
@@ -1,0 +1,6 @@
+package com.example.dispositivomobil.models
+
+data class Note(
+    val title: String,
+    val description: String
+)
